refactor background into its own file
