+++ bliss/datasets/galsim_galaxies.py
@@ -8,6 +8,8 @@ import torch
 from torch import Tensor
 from torch.utils.data import DataLoader, Dataset
 
+from bliss.datasets.background import ConstantBackground
+
 
 def load_psf_from_file(psf_image_file: str, pixel_scale: float):
     """Return normalized PSF galsim.GSObject from numpy psf_file."""
@@ -89,39 +91,6 @@ class ToyGaussian(pl.LightningDataModule, Dataset):
         image += noise
 
         return {"images": image, "background": self.background}
-
-    def __len__(self):
-        return self.batch_size * self.n_batches
-
-    def train_dataloader(self):
-        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def val_dataloader(self):
-        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def test_dataloader(self):
-        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
-
-
-class SDSSGalaxies(pl.LightningDataModule, Dataset):
-    def __init__(
-        self,
-        prior,
-        decoder,
-        num_workers,
-        batch_size,
-        n_batches,
-    ):
-        super().__init__()
-        self.prior = prior
-        self.decoder = decoder
-        self.n_batches = n_batches
-        self.batch_size = batch_size
-        self.num_workers = num_workers
-
-    def __getitem__(self, idx):
-        galaxy_params = self.prior.sample(1, "cpu")
-        return self.decoder.render_galaxy(galaxy_params[0])
 
     def __len__(self):
         return self.batch_size * self.n_batches
@@ -222,6 +191,7 @@ class GalsimGalaxyPrior:
 
     def _draw_pareto_flux(self, n_samples=1) -> Tensor:
         # draw pareto conditioned on being less than f_max
+        assert self.alpha is not None
         u_max = 1 - (self.min_flux / self.max_flux) ** self.alpha
         uniform_samples = torch.rand(n_samples) * u_max
         return self.min_flux / (1.0 - uniform_samples) ** (1 / self.alpha)
@@ -237,7 +207,6 @@ class GalsimGalaxyDecoder:
         self,
         slen,
         n_bands,
-        background,
         pixel_scale,
         psf_image_file: str,
     ) -> None:
@@ -245,8 +214,6 @@ class GalsimGalaxyDecoder:
         self.slen = slen
         assert n_bands == 1, "Only 1 band is supported"
         self.n_bands = 1
-        self.background = torch.zeros((self.n_bands, self.slen, self.slen), dtype=torch.float32)
-        self.background[...] = background
         self.pixel_scale = pixel_scale
         self.psf = load_psf_from_file(psf_image_file, self.pixel_scale)
 
@@ -289,8 +256,48 @@ class GalsimGalaxyDecoder:
             nx=self.slen, ny=self.slen, method="auto", scale=self.pixel_scale
         )
         image = torch.from_numpy(image.array).reshape(1, self.slen, self.slen)
-        noiseless = image.clone()
-        # add noise and background.
-        image += self.background.mean()
-        image += image.sqrt() * torch.randn(*image.shape)
-        return {"images": image, "background": self.background, "noiseless": noiseless}
+        return image
+        # noiseless = image.clone()
+        # # add noise and background.
+        # image += self.background.mean()
+        # image += image.sqrt() * torch.randn(*image.shape)
+        # return {"images": image, "background": self.background, "noiseless": noiseless}
+
+
+class SDSSGalaxies(pl.LightningDataModule, Dataset):
+    def __init__(
+        self,
+        prior: GalsimGalaxyPrior,
+        decoder: GalsimGalaxyDecoder,
+        background: ConstantBackground,
+        num_workers: int,
+        batch_size: int,
+        n_batches: int,
+    ):
+        super().__init__()
+        self.prior = prior
+        self.decoder = decoder
+        self.n_batches = n_batches
+        self.batch_size = batch_size
+        self.num_workers = num_workers
+        self.background = background
+
+    def __getitem__(self, idx):
+        galaxy_params = self.prior.sample(1, "cpu")
+        galaxy_image = self.decoder.render_galaxy(galaxy_params[0])
+        background = self.background.sample(galaxy_image.shape)
+        galaxy_with_noise = galaxy_image + background
+        galaxy_with_noise += galaxy_with_noise.sqrt() * torch.randn(*galaxy_with_noise)
+        return {"images": galaxy_with_noise, "background": background, "noiseless": galaxy_image}
+
+    def __len__(self):
+        return self.batch_size * self.n_batches
+
+    def train_dataloader(self):
+        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
+
+    def val_dataloader(self):
+        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
+
+    def test_dataloader(self):
+        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
