+++ bliss/datasets/simulated.py
@@ -1,16 +1,14 @@
 import warnings
 from typing import Dict, Tuple, Union
 
-import numpy as np
 import pytorch_lightning as pl
 import torch
-from einops import rearrange
-from torch import Tensor, nn
+from torch import Tensor
 from torch.utils.data import DataLoader, Dataset, IterableDataset
 
 from bliss.catalog import TileCatalog
+from bliss.datasets.background import ConstantBackground, SimulatedSDSSBackground
 from bliss.datasets.galsim_galaxies import GalsimGalaxyPrior
-from bliss.datasets.sdss import SloanDigitalSkySurvey
 from bliss.models.decoder import ImageDecoder
 from bliss.models.prior import ImagePrior
 
@@ -18,43 +16,6 @@ from bliss.models.prior import ImagePrior
 warnings.filterwarnings(
     "ignore", ".*does not have many workers which may be a bottleneck.*", UserWarning
 )
-
-
-class ConstantBackground(nn.Module):
-    def __init__(self, background):
-        super().__init__()
-        background = torch.tensor(background)
-        background = rearrange(background, "c -> 1 c 1 1")
-        self.register_buffer("background", background, persistent=False)
-
-    def sample(self, shape) -> Tensor:
-        batch_size, c, hlen, wlen = shape
-        return self.background.expand(batch_size, c, hlen, wlen)
-
-
-class SimulatedSDSSBackground(nn.Module):
-    def __init__(self, sdss_dir, run, camcol, field, bands):
-        super().__init__()
-        sdss_data = SloanDigitalSkySurvey(
-            sdss_dir=sdss_dir,
-            run=run,
-            camcol=camcol,
-            fields=(field,),
-            bands=bands,
-        )
-        background = torch.from_numpy(sdss_data[0]["background"])
-        background = rearrange(background, "c h w -> 1 c h w", c=len(bands))
-        self.register_buffer("background", background, persistent=False)
-        self.height, self.width = self.background.shape[-2:]
-
-    def sample(self, shape) -> Tensor:
-        batch_size, c, hlen, wlen = shape
-        assert self.background.shape[1] == c
-        h_diff, w_diff = self.height - hlen, self.width - wlen
-        h = 0 if h_diff == 0 else np.random.randint(h_diff)
-        w = 0 if w_diff == 0 else np.random.randint(w_diff)
-        bg = self.background[:, :, h : (h + hlen), w : (w + wlen)]
-        return bg.expand(batch_size, -1, -1, -1)
 
 
 class SimulatedDataset(pl.LightningDataModule, IterableDataset):
